requirements.txt
spotipy==2.23.0
yt-dlp>=2025.9.5
fastapi>=0.115.2
uvicorn[standard]>=0.24.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
python-multipart>=0.0.18
websockets==12.0
requests==2.31.0
colorama==0.4.6
starlette>=0.40.0
aiosqlite>=0.19.0
mutagen>=1.47.0
librosa>=0.10.1
scipy==1.10.1
